cleaned up server code, Object.hpp now uses dvec and doubles instead of floats

+++ src/Object.cpp
@@ -6,53 +6,53 @@
 
 #include <chrono>
 
-void Object::SetSize(glm::vec2 size) {
+void Object::SetSize(glm::dvec2 size) {
     m_size.x = size.x;
     m_size.y = size.y;
 }
 
-void Object::SetPosition(glm::vec2 position) {
+void Object::SetPosition(glm::dvec2 position) {
     m_position.x = position.x - this->GetSize().x * 0.5;
     m_position.y = position.y - this->GetSize().y * 0.5;
 }
 
-void Object::SetRotation(float rotation) {
+void Object::SetRotation(double rotation) {
     m_rotation = rotation;
 }
 
-void Object::SetSpeed(glm::vec2 speed) {
+void Object::SetSpeed(glm::dvec2 speed) {
     m_speed = speed;
 }
 
-void Object::SetAcceleration(glm::vec2 acceleration) {
+void Object::SetAcceleration(glm::dvec2 acceleration) {
     m_acceleration = acceleration;
 }
 
-glm::vec2 Object::GetSize() const {
+glm::dvec2 Object::GetSize() const {
     return m_size;
 }
 
-glm::vec2 Object::GetPosition() const {
-    return glm::vec2( m_position.x + 0.5 * m_size.x, m_position.y + 0.5 * m_size.y );
+glm::dvec2 Object::GetPosition() const {
+    return glm::dvec2( m_position.x + 0.5 * m_size.x, m_position.y + 0.5 * m_size.y );
 }
 
-float Object::GetRotation() const {
+double Object::GetRotation() const {
     return m_rotation;
 }
 
-glm::vec2 Object::GetSpeed() const {
+glm::dvec2 Object::GetSpeed() const {
     return m_speed;
 }
 
-glm::vec2 Object::GetAcceleration() const {
+glm::dvec2 Object::GetAcceleration() const {
     return m_acceleration;
 }
 
-void Object::Accelerate(glm::vec2 acceleration) {
+void Object::Accelerate(glm::dvec2 acceleration) {
     m_speed += acceleration;
 }
 
-void Object::Process(float dt) {
+void Object::Process(double dt) {
 	#ifndef SERVER
 	if(dt == 0)
 		dt = GameState::deltaTime;
@@ -62,11 +62,11 @@ void Object::Process(float dt) {
 	m_rotation += m_rotation_speed * dt;
 }
 
-void Object::SetRotationSpeed( float rotation_speed ) {
+void Object::SetRotationSpeed( double rotation_speed ) {
 	m_rotation_speed = rotation_speed;
 }
 
-void Object::InterpolateToState(Object &obj, float interpolation) {
+void Object::InterpolateToState(Object &obj, double interpolation) {
 	m_speed += (obj.m_speed - m_speed) * interpolation;
 	m_position += (obj.m_position - m_position) * interpolation;
 	m_rotation += (obj.m_rotation - m_rotation) * interpolation;
@@ -83,10 +83,10 @@ void Object::UpdateTicks() {
 	m_ticks = std::chrono::duration_cast<std::chrono::microseconds>(std::chrono::high_resolution_clock::now() - first_tick).count();
 }
 
-glm::vec2 Object::local_to_world_coord(const glm::vec2& local_coord) {
-	glm::vec2 global_coord;
-	float angle = m_rotation * 3.141592f / 180.0f;
-	float s = sin(angle), c = cos(angle);
+glm::dvec2 Object::local_to_world_coord(const glm::dvec2& local_coord) {
+	glm::dvec2 global_coord;
+	double angle = m_rotation * 3.141592 / 180.0;
+	double s = sin(angle), c = cos(angle);
 	// A*cos(x + r) = (A*cos(x))cos(r) - (A*sin(x))sin(r)
 	// A*sin(x + r) = (A*sin(x))cos(r) + (A*cos(x))sin(r)
