+++ src/Object.hpp
@@ -6,30 +6,30 @@
 
 class Object {
 	public:
-		Object(glm::vec2 size = {0,0}, glm::vec2 position = {0,0}, float rotation = 0, glm::vec2 speed = {0,0}) : 
+		Object(glm::vec2 size = {0,0}, glm::vec2 position = {0,0}, double rotation = 0, glm::vec2 speed = {0,0}) : 
 			m_size(size), m_position(position), m_rotation(rotation), m_speed(speed), m_rotation_speed(0) {}
 		~Object() {}
 
-		void SetSize(glm::vec2 size);
-		void SetPosition(glm::vec2 position);
-		void SetRotation(float rotation);
-		void SetSpeed(glm::vec2 speed);
-		void SetAcceleration(glm::vec2 acceleration);
+		void SetSize(glm::dvec2 size);
+		void SetPosition(glm::dvec2 position);
+		void SetRotation(double rotation);
+		void SetSpeed(glm::dvec2 speed);
+		void SetAcceleration(glm::dvec2 acceleration);
 		void CopyObjectState(Object &obj);
-		void InterpolateToState(Object &obj, float interpolation);
+		void InterpolateToState(Object &obj, double interpolation);
 		
-		glm::vec2 GetSize() const;
-		glm::vec2 GetPosition() const;
-		float GetRotation() const;
-		glm::vec2 GetSpeed() const;
-		glm::vec2 GetAcceleration() const;
-		void SetRotationSpeed( float rotation_speed );
+		glm::dvec2 GetSize() const;
+		glm::dvec2 GetPosition() const;
+		double GetRotation() const;
+		glm::dvec2 GetSpeed() const;
+		glm::dvec2 GetAcceleration() const;
+		void SetRotationSpeed( double rotation_speed );
 		
-		void Accelerate(glm::vec2 acceleration);
+		void Accelerate(glm::dvec2 acceleration);
 
 		void Draw() {};
 		
-		void Process(float dt = 0);
+		void Process(double dt = 0);
 		unsigned int GetId() { return id; }
 		void SetId(unsigned int id) { this->id = id; }
 		unsigned int GetTicks() { return m_ticks; }
@@ -40,16 +40,16 @@ class Object {
 	protected:
 		uint32_t id;
 		
-		glm::vec2 m_size;
-		glm::vec2 m_position;
-		float m_rotation;
-		float m_rotation_speed;
-		glm::vec2 m_speed;
+		glm::dvec2 m_size;
+		glm::dvec2 m_position;
+		double m_rotation;
+		double m_rotation_speed;
+		glm::dvec2 m_speed;
 		uint32_t m_ticks;
 		
-		glm::vec2 m_acceleration;
+		glm::dvec2 m_acceleration;
 		
-		glm::vec2 local_to_world_coord(const glm::vec2& local_coord);
+		glm::dvec2 local_to_world_coord(const glm::dvec2& local_coord);
 		
 	private:
