+++ src/Ship.hpp
@@ -14,12 +14,12 @@ class Ship : public SolidObject {
 			std::string name;
 			GLuint texture;
 			GLuint skin;
-			float mass;
-			float armor;
+			double mass;
+			double armor;
 			Sprite sprite;
 		};
 
-		Ship(glm::vec2 position, float rotation, const Chassis& chassis);
+		Ship(glm::dvec2 position, double rotation, const Chassis& chassis);
 		~Ship();
 		
 		void Process();
@@ -32,13 +32,13 @@ class Ship : public SolidObject {
 		
 	private:
 		Chassis m_chassis;
-		float m_rotation_speed_coefficient;
+		double m_rotation_speed_coefficient;
 		int m_max_distance_acceleration;
-		float m_downshift_coefficient;
-		float m_max_speed_coefficient;
-		float m_acceleration_speed_coefficient;
-		float m_brake_coefficient;
-		float m_engine_propulsion_coefficient;
+		double m_downshift_coefficient;
+		double m_max_speed_coefficient;
+		double m_acceleration_speed_coefficient;
+		double m_brake_coefficient;
+		double m_engine_propulsion_coefficient;
 		bool m_stabilizers_on;
 		Sprite m_engine_propulsion;
 };
